+++ src/NeuralNet/Optimizers/Adam.php
@@ -100,7 +100,7 @@ class Adam implements Optimizer, Adaptive
     {
         $class = get_class($param->param());
 
-        if ($class === false) {
+        if (!$class) {
             throw new RuntimeException('Could not locate parameter class.');
         }