+++ src/NeuralNet/Optimizers/RMSProp.php
@@ -91,7 +91,7 @@ class RMSProp implements Optimizer, Adaptive
     {
         $class = get_class($param->param());
 
-        if ($class === false) {
+        if (!$class) {
             throw new RuntimeException('Could not locate parameter class.');
         }